+++ CHANGES.rst
@@ -3,7 +3,8 @@ Changes
 
 * vHEAD:
 
-  - ...
+  - Fix set_tokens() bug in Python3; make sure json.loads() gets an
+    unistr. Reported by @Imperatus, @LordGaav. Closes #9 and #10.
 
 * v0.2.4: